LayoutTests/http/wpt/web-animations/interfaces/AnimationEffectTiming/iterationStart-expected.txt


FAIL Test default value assert_equals: expected (number) 0 but got (undefined) undefined
FAIL Test that changing the iterationStart affects computed timing when backwards-filling anim.effect.getComputedTiming is not a function. (In 'anim.effect.getComputedTiming()', 'anim.effect.getComputedTiming' is undefined)
FAIL Test that changing the iterationStart affects computed timing during the active phase anim.effect.getComputedTiming is not a function. (In 'anim.effect.getComputedTiming()', 'anim.effect.getComputedTiming' is undefined)
FAIL Test that changing the iterationStart affects computed timing when forwards-filling anim.effect.getComputedTiming is not a function. (In 'anim.effect.getComputedTiming()', 'anim.effect.getComputedTiming' is undefined)
FAIL Using invalid values should throw TypeError assert_throws: setting -1 function "function () {
      anim.effect.timing.iterationStart = invalid;
    }" did not throw
